make safe_extract() to common utility.
